requirements.txt
fastapi
pydantic
uvicorn
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
